feat(log): add a log adapter for retryablehttp.LeveledLogger
