+++ pkg/log/adapters/adapters_test.go
@@ -46,3 +46,24 @@ func ExampleNewLogrLogger() {
 	// {"@timestamp":"2021-12-21T14:19:20.0424249-08:00","app.version":"testing","b":"hello, world!","c":1,"level":"INFO","message":"info!!"}
 	// {"@timestamp":"2021-12-21T14:19:20.0424249-08:00","app.version":"testing","b":"hello, world!","c":1,"error.error":"bad thing","error.kind":"error","error.message":"bad thing","level":"ERROR","message":"end of the world!"}
 }
+
+func ExampleNewLeveledLogger() {
+	logs := logtest.NewLogRecorder(nil)
+	defer logs.Close()
+
+	logger := adapters.NewLeveledLogger(context.Background())
+
+	logger.Info("hello, info", "a", 1)
+	logger.Debug("hello, debug", "a", 1)
+	logger.Error("hello, error", "a", 1)
+	logger.Warn("hello, warn", "a", 1)
+
+	printEntries(logs.Entries())
+
+	//nolint:lll // Why: testing output
+	// Output:
+	// {"@timestamp":"2021-12-21T14:19:20.0424249-08:00","a":1,"app.version":"testing","level":"INFO","message":"hello, info"}
+	// {"@timestamp":"2021-12-21T14:19:20.0424249-08:00","a":1,"app.version":"testing","level":"DEBUG","message":"hello, debug"}
+	// {"@timestamp":"2021-12-21T14:19:20.0424249-08:00","a":1,"app.version":"testing","level":"ERROR","message":"hello, error"}
+	// {"@timestamp":"2021-12-21T14:19:20.0424249-08:00","a":1,"app.version":"testing","level":"WARN","message":"hello, warn"}
+}
